requirements.txt
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.3.2
Django==2.1.5
docopt==0.6.2
idna==3.6
olefile==0.46
Pillow==5.4.1
pytz @ file:///C:/b/abs_22fofvpn1x/croot/pytz_1671698059864/work
requests==2.31.0
urllib3==2.0.7
wincertstore==0.2
yarg==0.1.9
django_taggit
